backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI and LLM
langchain==0.0.350
langchain-openai==0.0.2
openai>=1.6.1

# Vector database (optional, for pgvector)
pgvector==0.2.4

# Data validation
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# HTTP client
httpx==0.25.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production dependencies
gunicorn==21.2.0